+++ README.md
@@ -16,12 +16,12 @@ github.com/AlDanial/cloc v 1.94  T=0.05 s (949.3 files/s, 62349.4 lines/s)
 -------------------------------------------------------------------------------
 Language                     files          blank        comment           code
 -------------------------------------------------------------------------------
-C                               37            508            655           1562
-C/C++ Header                    10             68            104            283
-Assembly                         3              9             31             91
+C                               36            499            632           1547
+C/C++ Header                    10             70            105            298
+Assembly                         3              9             29             91
 make                             1             14              3             64
 -------------------------------------------------------------------------------
-SUM:                            51            599            793           2000 (exactly 2000!)
+SUM:                            50            592            769           2000 (exactly 2000!)
 -------------------------------------------------------------------------------
 ```
